backend-image/requirements.txt
fastapi
pydantic
gunicorn
uvicorn
azure-core
requests
openai
python-dotenv